# Performance optimization for non-stream requests