@ikhoon 2021-02-07
| Server | TPS | Latency |
| --- | --- | --- |
| Plain Armeria | 118,521 | 14.59ms |
| http4s-blaze | **105,482** | 21.76ms |
| http4s-armeria | 83,888 | **14.44ms** |

Note: 
- The CPU hits 100% during the benchmarks on all servers.
- `http4s-armeria` shows better latency.
- `http4s-blaze` shows better TPS.
- `Armeria` is better than both of them.